update HMS - dockerized, added prod config, billing fixes

+++ hms-backend/src/main/java/com/example/hms/dto/billing_dto/BillItemDto.java
@@ -6,6 +6,8 @@ import lombok.Builder;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @Data
 @Builder
 @NoArgsConstructor
@@ -15,6 +17,6 @@ public class BillItemDto {
     private String itemName;
     private ItemTypes itemType;
     private Integer quantity;
-    private Double unitPrice;
-    private Double lineTotal;
+    private BigDecimal unitPrice;
+    private BigDecimal lineTotal;
 }
