+++ hms-backend/src/main/java/com/example/hms/model/billingModel/BillItem.java
@@ -6,6 +6,8 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
@@ -25,9 +27,9 @@ public class BillItem {
 
     private Integer quantity;
 
-    private Double unitPrice;
+    private BigDecimal unitPrice;
 
-    private Double lineTotal;
+    private BigDecimal lineTotal;
 
     @ManyToOne
     @JoinColumn(name = "bill_id")
